third_party/WebKit/LayoutTests/fast/pagination/caret-range-outside-paged-x-expected.txt

Test that hitting areas outside a paged container takes us to the right text content on some page.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
To the left of page 1:
PASS document.caretRangeFromPoint(0,20).startOffset is 0
PASS document.caretRangeFromPoint(0,60).startOffset is 0
PASS document.caretRangeFromPoint(0,100).startOffset is 4
Above page 1:
PASS document.caretRangeFromPoint(80,0).startOffset is 0
PASS document.caretRangeFromPoint(200,0).startOffset is 0
PASS document.caretRangeFromPoint(299,0).startOffset is 0
Below page 1:
PASS document.caretRangeFromPoint(80,200).startOffset is 12
PASS document.caretRangeFromPoint(200,200).startOffset is 12
PASS document.caretRangeFromPoint(299,200).startOffset is 12
Scroll to page 2
Above page 2:
PASS document.caretRangeFromPoint(200,0).startOffset is 12
PASS document.caretRangeFromPoint(299,0).startOffset is 12
Below page 2:
PASS document.caretRangeFromPoint(200,200).startOffset is 24
PASS document.caretRangeFromPoint(299,200).startOffset is 24
Scroll to page 3
Above page 3:
PASS document.caretRangeFromPoint(200,0).startOffset is 24
PASS document.caretRangeFromPoint(299,0).startOffset is 24
To the right of page 3:
PASS document.caretRangeFromPoint(400,20).startOffset is 24
PASS document.caretRangeFromPoint(400,60).startOffset is 27
PASS document.caretRangeFromPoint(400,100).startOffset is 31
PASS document.caretRangeFromPoint(400,140).startOffset is 35
xxx xxx xxx xxx xxx xxx xxx xxx xxx
